pyproject.toml
[project]
name = "newsapi-gmail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.174.0",
    "google-auth-oauthlib>=1.2.2",
    "requests>=2.32.4",
    "beautifulsoup4>=4.12.3",
    "google-generativeai>=0.7.2",
    "python-dotenv>=1.0.1",
]
